refactor unittest python adapter code to match pytest version
