+++ Python/Product/TestAdapter.Executor/PythonFiles/testing_tools/adapter/unittest/_cli.py
@@ -1,0 +1,18 @@
+﻿# Copyright (c) Microsoft Corporation. All rights reserved.
+# Licensed under the MIT License.
+
+from __future__ import absolute_import
+
+from ..errors import UnsupportedCommandError
+
+
+
+def add_unittest_subparser(cmd, name, parent):
+    """Add a new subparser to the given parent and add args to it."""
+    parser = parent.add_parser(name)
+    if cmd == 'discover':
+        # For now we don't have any tool-specific CLI options to add.
+        pass
+    else:
+        raise UnsupportedCommandError(cmd)
+    return parser
