+++ Python/Product/TestAdapter.Executor/UnitTest/TestExecutorUnitTest.cs
@@ -21,7 +21,6 @@ using System.Diagnostics.CodeAnalysis;
 using System.IO;
 using System.Linq;
 using System.Net;
-using System.Net.NetworkInformation;
 using System.Net.Sockets;
 using System.Runtime.InteropServices;
 using System.Security.Cryptography;
@@ -143,7 +142,7 @@ namespace Microsoft.PythonTools.TestAdapter {
 
         public void RunTests(IEnumerable<TestCase> tests, IRunContext runContext, IFrameworkHandle frameworkHandle) {
 
-            MessageBox.Show("Hello1: " + Process.GetCurrentProcess().Id);
+            //MessageBox.Show("Hello1: " + Process.GetCurrentProcess().Id);
 
             if (tests == null) {
                 throw new ArgumentNullException(nameof(tests));
